Update download_file util

+++ toai/utils/download_file.py
@@ -1,9 +1,15 @@
-import requests
+import os
 import shutil
+from pathlib import Path
+from typing import Union
+
+import requests
 
 
-def download_file(url: str):
-    local_filename = url.split("/")[-1]
+def download_file(url: str, path: Union[Path, str] = Path("."), override: bool = False):
+    local_filename = f"{str(path)}/{url.split('/')[-1]}"
+    if os.path.exists(local_filename) and not override:
+        raise FileExistsError(f"File exists: {local_filename}")
     with requests.get(url, stream=True) as r:
         with open(local_filename, "wb") as f:
             shutil.copyfileobj(r.raw, f)
